Binary search tree
This builds an interface and then takes in either integer or fractional input and builds a binary search tree based on their values. The binary search tree is generic so that a new data type of fraction can be built. This type fraction is defined specifically for the compare to so that it will be correctly placed in the tree.